refactor(http): move helper subroutines from SocketHTTPServer.c to server/
Relocate internal helper functions to their proper homes:

- Timeout helpers → SocketHTTPServer-connections.c
  - check_global_lifetime_timeout, check_tls_handshake_timeout
  - check_request_timeout, check_response_timeout, check_http2_timeout
  - server_check_connection_timeout

- Event processing → SocketHTTPServer-connections.c
  - server_process_tls_handshake, server_process_websocket
  - server_process_client_event

- HTTP/2 processing → SocketHTTPServer-h2.c
  - server_process_http2, server_try_http2_prior_knowledge
  - validate_rfc8441_websocket_upgrade, prepare_h2_websocket_response
  - setup_ws_over_h2_streaming

- Rate limiting/validation → SocketHTTPServer-http1.c
  - server_check_rate_limit, server_run_validator
  - server_run_validator_early

The main SocketHTTPServer.c now contains only public API implementations
and find_rate_limiter which is shared across files.

Fixes #2665
