+++ src/http/SocketHTTPServer.c
@@ -7,33 +7,22 @@
 /* SocketHTTPServer.c - HTTP/1.1 and HTTP/2 server with TLS, rate limiting, and
  * connection pooling */
 
-#include <arpa/inet.h>
 #include <assert.h>
-#include <errno.h>
-#include <pthread.h>
 #include <stdlib.h>
 #include <string.h>
-#include <sys/socket.h>
 
 #include "core/Arena.h"
-#include "core/SocketIPTracker.h"
-#include "core/SocketMetrics.h"
 #include "core/SocketRateLimit.h"
-#include "core/SocketCrypto.h"
 #include "core/SocketUtil.h"
 #include "http/SocketHTTP.h"
 #include "http/SocketHTTP1.h"
-#include "http/SocketHTTP2-private.h"
-#include "http/SocketHTTPServer-http1.h"
+#include "http/SocketHTTP2.h"
 #include "http/SocketHTTPServer-private.h"
 #include "http/SocketHTTPServer.h"
 #include "poll/SocketPoll.h"
 #include "socket/Socket.h"
 #include "socket/SocketBuf.h"
 #include "socket/SocketWS.h"
-#if SOCKET_HAS_TLS
-#include "tls/SocketTLS.h"
-#endif
 
 #undef SOCKET_LOG_COMPONENT
 #define SOCKET_LOG_COMPONENT "HTTPServer"
@@ -73,650 +62,6 @@ find_rate_limiter (SocketHTTPServer_T server, const char *path)
   if (best != NULL)
     return best->limiter;
   return server->global_rate_limiter;
-}
-
-#if SOCKET_HAS_TLS
-static int
-server_process_tls_handshake (SocketHTTPServer_T server,
-                              ServerConnection *conn,
-                              unsigned events)
-{
-  volatile TLSHandshakeState hs = TLS_HANDSHAKE_NOT_STARTED;
-
-  (void)events;
-
-  assert (server != NULL);
-  assert (conn != NULL);
-
-  TRY
-  {
-    hs = SocketTLS_handshake (conn->socket);
-  }
-  EXCEPT (SocketTLS_HandshakeFailed)
-  {
-    conn->state = CONN_STATE_CLOSED;
-    return -1;
-  }
-  EXCEPT (SocketTLS_VerifyFailed)
-  {
-    conn->state = CONN_STATE_CLOSED;
-    return -1;
-  }
-  EXCEPT (SocketTLS_Failed)
-  {
-    conn->state = CONN_STATE_CLOSED;
-    return -1;
-  }
-  END_TRY;
-
-  if (hs == TLS_HANDSHAKE_COMPLETE)
-    {
-      conn->tls_handshake_done = 1;
-
-      /* Decide protocol by ALPN. If not negotiated, fall back to HTTP/1.1. */
-      const char *alpn = SocketTLS_get_alpn_selected (conn->socket);
-      if (alpn != NULL && strcmp (alpn, "h2") == 0
-          && server->config.max_version >= HTTP_VERSION_2)
-        {
-          /* Enable HTTP/2 - use early return to reduce nesting. */
-          if (server_http2_enable (server, conn) < 0)
-            {
-              conn->state = CONN_STATE_CLOSED;
-              return -1;
-            }
-          conn->is_http2 = 1;
-          conn->state = CONN_STATE_HTTP2;
-          SocketPoll_mod (
-              server->poll, conn->socket, POLL_READ | POLL_WRITE, conn);
-        }
-      else
-        {
-          conn->is_http2 = 0;
-          conn->state = CONN_STATE_READING_REQUEST;
-          SocketPoll_mod (server->poll, conn->socket, POLL_READ, conn);
-        }
-
-      return 0;
-    }
-
-  /* Continue handshake: narrow poll interest to avoid busy loops. */
-  if (hs == TLS_HANDSHAKE_WANT_READ)
-    SocketPoll_mod (server->poll, conn->socket, POLL_READ, conn);
-  else if (hs == TLS_HANDSHAKE_WANT_WRITE)
-    SocketPoll_mod (server->poll, conn->socket, POLL_WRITE, conn);
-  else
-    SocketPoll_mod (server->poll, conn->socket, POLL_READ | POLL_WRITE, conn);
-
-  return 0;
-}
-#endif
-
-static int
-server_process_http2 (SocketHTTPServer_T server,
-                      ServerConnection *conn,
-                      unsigned events)
-{
-  volatile int r = 0;
-  volatile int f = 0;
-  volatile int stream_error = 0;
-
-  assert (server != NULL);
-  assert (conn != NULL);
-
-  if (conn->http2_conn == NULL)
-    {
-      if (server_http2_enable (server, conn) < 0)
-        return -1;
-    }
-
-  TRY
-  {
-    r = SocketHTTP2_Conn_process (conn->http2_conn, events);
-  }
-  EXCEPT (SocketHTTP2_ProtocolError)
-  {
-    return -1;
-  }
-  EXCEPT (SocketHTTP2_FlowControlError)
-  {
-    return -1;
-  }
-  EXCEPT (SocketHTTP2_StreamError)
-  {
-    /* Stream-level error: non-fatal for the connection (RFC 9113).
-     * The core resets the offending stream; other streams may continue. */
-    if (Except_frame.exception != NULL
-        && Except_frame.exception->reason != NULL)
-      SOCKET_LOG_WARN_MSG ("HTTP/2 stream error: %s",
-                           Except_frame.exception->reason);
-    stream_error = 1;
-    r = 0;
-  }
-  EXCEPT (Socket_Failed)
-  {
-    return -1;
-  }
-  END_TRY;
-
-  if (!stream_error && r < 0)
-    return -1;
-
-  TRY
-  {
-    f = SocketHTTP2_Conn_flush (conn->http2_conn);
-  }
-  EXCEPT (Socket_Failed)
-  {
-    return -1;
-  }
-  EXCEPT (SocketHTTP2_FlowControlError)
-  {
-    return -1;
-  }
-  END_TRY;
-
-  if (SocketHTTP2_Conn_is_closed (conn->http2_conn))
-    return -1;
-
-  if (f == 1)
-    SocketPoll_mod (server->poll, conn->socket, POLL_READ | POLL_WRITE, conn);
-  else
-    SocketPoll_mod (server->poll, conn->socket, POLL_READ, conn);
-
-  return 0;
-}
-
-static int
-server_process_websocket (SocketHTTPServer_T server,
-                          ServerConnection *conn,
-                          unsigned events)
-{
-  volatile int status = 0;
-
-  assert (server != NULL);
-  assert (conn != NULL);
-  assert (conn->websocket != NULL);
-
-  TRY
-  {
-    status = SocketWS_process (conn->websocket, events);
-  }
-  EXCEPT (SocketWS_Failed)
-  {
-    return -1;
-  }
-  EXCEPT (SocketWS_ProtocolError)
-  {
-    return -1;
-  }
-  EXCEPT (SocketWS_Closed)
-  {
-    return -1;
-  }
-  END_TRY;
-
-  if (status < 0)
-    return -1;
-
-  /* Process received messages via callback */
-  while (SocketWS_recv_available (conn->websocket) > 0)
-    {
-      SocketWS_Message msg;
-      int result = SocketWS_recv_message (conn->websocket, &msg);
-      if (result > 0 && conn->ws_callback != NULL)
-        {
-          int is_final
-              = (SocketWS_state (conn->websocket) == WS_STATE_CLOSING
-                 || SocketWS_state (conn->websocket) == WS_STATE_CLOSED)
-                    ? 1
-                    : 0;
-          conn->ws_callback (
-              NULL, msg.data, msg.len, is_final, conn->ws_callback_userdata);
-          free (msg.data);
-        }
-      else if (result <= 0)
-        {
-          return -1;
-        }
-    }
-
-  /* Check if connection closed */
-  if (SocketWS_state (conn->websocket) == WS_STATE_CLOSED)
-    return -1;
-
-  /* Update poll events */
-  unsigned ws_events = SocketWS_poll_events (conn->websocket);
-  SocketPoll_mod (server->poll, conn->socket, ws_events, conn);
-
-  return 0;
-}
-
-static int
-server_try_http2_prior_knowledge (SocketHTTPServer_T server,
-                                  ServerConnection *conn,
-                                  unsigned events)
-{
-  unsigned char preface[HTTP2_PREFACE_SIZE];
-  int fd;
-  ssize_t n;
-
-  assert (server != NULL);
-  assert (conn != NULL);
-
-  if (!(events & POLL_READ))
-    return 0;
-  if (server->config.tls_context != NULL)
-    return 0;
-  if (server->config.max_version < HTTP_VERSION_2)
-    return 0;
-  if (conn->state != CONN_STATE_READING_REQUEST)
-    return 0;
-  if (conn->http2_conn != NULL || conn->is_http2)
-    return 0;
-
-  fd = Socket_fd (conn->socket);
-  n = recv (fd, preface, sizeof (preface), MSG_PEEK | MSG_DONTWAIT);
-  if (n < 0)
-    {
-      if (errno == EAGAIN || errno == EWOULDBLOCK)
-        return 0;
-      return 0;
-    }
-  if ((size_t)n < sizeof (preface))
-    return 0;
-
-  if (memcmp (preface, HTTP2_CLIENT_PREFACE, HTTP2_PREFACE_SIZE) != 0)
-    return 0;
-
-  conn->is_http2 = 1;
-  conn->state = CONN_STATE_HTTP2;
-
-  if (server_http2_enable (server, conn) < 0)
-    {
-      conn->state = CONN_STATE_CLOSED;
-      return 0;
-    }
-
-  (void)server_process_http2 (server, conn, events);
-  return 1;
-}
-
-/* Check rate limit for request path. Returns 1 if allowed, 0 if rate limited
- * (sends 429) */
-int
-server_check_rate_limit (SocketHTTPServer_T server, ServerConnection *conn)
-{
-  SocketRateLimit_T limiter
-      = find_rate_limiter (server, conn->request ? conn->request->path : NULL);
-  if (limiter != NULL && !SocketRateLimit_try_acquire (limiter, 1))
-    {
-      SERVER_METRICS_INC (
-          server, SOCKET_CTR_HTTP_SERVER_RATE_LIMITED, rate_limited);
-      connection_send_error (server, conn, 429, "Too Many Requests");
-      return 0;
-    }
-  return 1;
-}
-
-/* Run validator callback. Returns 1 if allowed, 0 if rejected */
-static int
-server_run_validator_impl (SocketHTTPServer_T server, ServerConnection *conn)
-{
-  int reject_status = 0;
-  struct SocketHTTPServer_Request req_ctx;
-
-  if (server->validator == NULL)
-    return 1;
-
-  req_ctx.server = server;
-  req_ctx.conn = conn;
-  req_ctx.h2_stream = NULL;
-  req_ctx.arena = conn->arena;
-  req_ctx.start_time_ms = conn->request_start_ms;
-
-  if (!server->validator (&req_ctx, &reject_status, server->validator_userdata))
-    {
-      if (reject_status == 0)
-        reject_status = 403;
-      connection_send_error (server, conn, reject_status, "Request Rejected");
-      return 0;
-    }
-
-  return 1;
-}
-
-/**
- * server_run_validator - Run request validator callback
- * @server: HTTP server
- * @conn: Connection with parsed request
- *
- * Returns: 1 if allowed, 0 if rejected (sends error)
- */
-int
-server_run_validator (SocketHTTPServer_T server, ServerConnection *conn)
-{
-  return server_run_validator_impl (server, conn);
-}
-
-/* Run validator early (after headers, before body). Allows setting up body
- * streaming */
-int
-server_run_validator_early (SocketHTTPServer_T server, ServerConnection *conn)
-{
-  return server_run_validator_impl (server, conn);
-}
-
-
-/**
- * server_process_client_event - Process a single client event
- * @server: HTTP server
- * @conn: Client connection
- * @events: Event flags (POLL_READ, POLL_WRITE, etc.)
- *
- * Returns: 1 if request processed, 0 otherwise
- */
-int
-server_process_client_event (SocketHTTPServer_T server,
-                             ServerConnection *conn,
-                             unsigned events)
-{
-  int requests_processed = 0;
-
-  if (server_try_http2_prior_knowledge (server, conn, events))
-    return 0;
-
-  /* Handle disconnect/error events first */
-  if (events & (POLL_HANGUP | POLL_ERROR))
-    {
-      conn->state = CONN_STATE_CLOSED;
-      connection_close (server, conn);
-      return 0;
-    }
-
-  if (events & POLL_READ)
-    {
-      /* TLS handshake, HTTP/2, and WebSocket handle their own I/O.
-       * Only read into HTTP/1.1 buffer for request parsing states. */
-      if (conn->state != CONN_STATE_TLS_HANDSHAKE
-          && conn->state != CONN_STATE_HTTP2
-          && conn->state != CONN_STATE_WEBSOCKET)
-        connection_read (server, conn);
-    }
-
-  if (conn->state == CONN_STATE_TLS_HANDSHAKE)
-    {
-#if SOCKET_HAS_TLS
-      if (server_process_tls_handshake (server, conn, events) < 0)
-        conn->state = CONN_STATE_CLOSED;
-#else
-      conn->state = CONN_STATE_CLOSED;
-#endif
-    }
-
-  if (conn->state == CONN_STATE_HTTP2)
-    {
-      if (server_process_http2 (server, conn, events) < 0)
-        conn->state = CONN_STATE_CLOSED;
-      if (conn->state == CONN_STATE_CLOSED)
-        {
-          connection_close (server, conn);
-        }
-      return 0;
-    }
-
-  if (conn->state == CONN_STATE_WEBSOCKET)
-    {
-      if (server_process_websocket (server, conn, events) < 0)
-        conn->state = CONN_STATE_CLOSED;
-      if (conn->state == CONN_STATE_CLOSED)
-        {
-          connection_close (server, conn);
-        }
-      return 0;
-    }
-
-  if (conn->state == CONN_STATE_READING_REQUEST)
-    {
-      if (connection_parse_request (server, conn) == 1)
-        {
-          requests_processed = server_handle_parsed_request (server, conn);
-        }
-    }
-
-  /* Continue reading request body using centralized parser API */
-  if (conn->state == CONN_STATE_READING_BODY)
-    {
-      int result = server_process_body_reading (server, conn);
-      if (result > 0)
-        requests_processed = result;
-    }
-
-  if (conn->state == CONN_STATE_CLOSED)
-    {
-      connection_close (server, conn);
-    }
-
-  return requests_processed;
-}
-
-/**
- * check_global_lifetime_timeout - Check global connection lifetime limit
- * @server: HTTP server
- * @conn: Connection to check
- * @now: Current time in milliseconds
- *
- * SECURITY: Defense-in-depth against connections held indefinitely.
- * Applies to all states. Set to 0 to disable.
- *
- * Returns: 1 if timed out (connection closed), 0 otherwise
- */
-static int
-check_global_lifetime_timeout (SocketHTTPServer_T server,
-                               ServerConnection *conn,
-                               int64_t now)
-{
-  if (server->config.max_connection_lifetime_ms <= 0)
-    return 0;
-
-  int64_t connection_age_ms = now - conn->created_at_ms;
-  if (connection_age_ms > server->config.max_connection_lifetime_ms)
-    {
-      SOCKET_LOG_WARN_MSG (
-          "Connection lifetime exceeded (%lld ms > %d ms), closing connection",
-          (long long)connection_age_ms,
-          server->config.max_connection_lifetime_ms);
-      SERVER_METRICS_INC (
-          server, SOCKET_CTR_HTTP_SERVER_REQUESTS_TIMEOUT, requests_timeout);
-      connection_close (server, conn);
-      return 1;
-    }
-  return 0;
-}
-
-/**
- * check_tls_handshake_timeout - Check TLS handshake timeout
- * @server: HTTP server
- * @conn: Connection to check
- * @now: Current time in milliseconds
- *
- * SECURITY: Prevents slowloris attacks during TLS negotiation phase.
- *
- * Returns: 1 if timed out (connection closed), 0 otherwise
- */
-static int
-check_tls_handshake_timeout (SocketHTTPServer_T server,
-                             ServerConnection *conn,
-                             int64_t now)
-{
-  if (conn->state != CONN_STATE_TLS_HANDSHAKE)
-    return 0;
-  if (server->config.tls_handshake_timeout_ms <= 0)
-    return 0;
-
-  int64_t idle_ms = now - conn->last_activity_ms;
-  if (idle_ms > server->config.tls_handshake_timeout_ms)
-    {
-      SOCKET_LOG_WARN_MSG (
-          "TLS handshake timeout (%lld ms > %d ms), closing connection",
-          (long long)idle_ms,
-          server->config.tls_handshake_timeout_ms);
-      SERVER_METRICS_INC (
-          server, SOCKET_CTR_HTTP_SERVER_REQUESTS_TIMEOUT, requests_timeout);
-      connection_close (server, conn);
-      return 1;
-    }
-  return 0;
-}
-
-/**
- * check_request_timeout - Check request parsing and body reading timeouts
- * @server: HTTP server
- * @conn: Connection to check
- * @now: Current time in milliseconds
- *
- * SECURITY: Prevents slowloris attacks where headers/body are sent slowly.
- *
- * Returns: 1 if timed out (connection closed), 0 otherwise
- */
-static int
-check_request_timeout (SocketHTTPServer_T server,
-                       ServerConnection *conn,
-                       int64_t now)
-{
-  int64_t idle_ms = now - conn->last_activity_ms;
-
-  /* Keepalive timeout for idle connections */
-  if (conn->state == CONN_STATE_READING_REQUEST
-      && idle_ms > server->config.keepalive_timeout_ms)
-    {
-      SERVER_METRICS_INC (
-          server, SOCKET_CTR_HTTP_SERVER_REQUESTS_TIMEOUT, requests_timeout);
-      connection_close (server, conn);
-      return 1;
-    }
-
-  /* Header parsing timeout */
-  if (conn->state == CONN_STATE_READING_REQUEST && conn->request_start_ms > 0
-      && (now - conn->request_start_ms)
-             > server->config.request_read_timeout_ms)
-    {
-      SOCKET_LOG_WARN_MSG (
-          "Header parsing timeout (%lld ms > %d ms), closing connection",
-          (long long)(now - conn->request_start_ms),
-          server->config.request_read_timeout_ms);
-      SERVER_METRICS_INC (
-          server, SOCKET_CTR_HTTP_SERVER_REQUESTS_TIMEOUT, requests_timeout);
-      connection_close (server, conn);
-      return 1;
-    }
-
-  /* Request body reading timeout */
-  if (conn->state == CONN_STATE_READING_BODY && conn->request_start_ms > 0
-      && (now - conn->request_start_ms)
-             > server->config.request_read_timeout_ms)
-    {
-      SERVER_METRICS_INC (
-          server, SOCKET_CTR_HTTP_SERVER_REQUESTS_TIMEOUT, requests_timeout);
-      connection_close (server, conn);
-      return 1;
-    }
-
-  return 0;
-}
-
-/**
- * check_response_timeout - Check response write timeout
- * @server: HTTP server
- * @conn: Connection to check
- * @now: Current time in milliseconds
- *
- * Returns: 1 if timed out (connection closed), 0 otherwise
- */
-static int
-check_response_timeout (SocketHTTPServer_T server,
-                        ServerConnection *conn,
-                        int64_t now)
-{
-  if (conn->state != CONN_STATE_STREAMING_RESPONSE)
-    return 0;
-  if (conn->response_start_ms <= 0)
-    return 0;
-
-  if ((now - conn->response_start_ms)
-      > server->config.response_write_timeout_ms)
-    {
-      SERVER_METRICS_INC (
-          server, SOCKET_CTR_HTTP_SERVER_REQUESTS_TIMEOUT, requests_timeout);
-      connection_close (server, conn);
-      return 1;
-    }
-  return 0;
-}
-
-/**
- * check_http2_timeout - Check HTTP/2 idle connection timeout
- * @server: HTTP server
- * @conn: Connection to check
- * @now: Current time in milliseconds
- *
- * SECURITY: Prevents resource exhaustion from idle HTTP/2 connections.
- *
- * Returns: 1 if timed out (connection closed), 0 otherwise
- */
-static int
-check_http2_timeout (SocketHTTPServer_T server,
-                     ServerConnection *conn,
-                     int64_t now)
-{
-  if (conn->state != CONN_STATE_HTTP2)
-    return 0;
-
-  int64_t idle_ms = now - conn->last_activity_ms;
-  if (idle_ms > server->config.keepalive_timeout_ms)
-    {
-      SOCKET_LOG_WARN_MSG ("HTTP/2 connection idle timeout (%lld ms > %d ms), "
-                           "closing connection",
-                           (long long)idle_ms,
-                           server->config.keepalive_timeout_ms);
-      SERVER_METRICS_INC (
-          server, SOCKET_CTR_HTTP_SERVER_REQUESTS_TIMEOUT, requests_timeout);
-      connection_close (server, conn);
-      return 1;
-    }
-  return 0;
-}
-
-/**
- * server_check_connection_timeout - Check if connection has timed out
- * @server: HTTP server
- * @conn: Connection to check
- * @now: Current time in milliseconds
- *
- * SECURITY: Enhanced timeout enforcement to prevent Slowloris attacks
- * - TLS handshake timeout (CONN_STATE_TLS_HANDSHAKE)
- * - HTTP/2 idle connection timeout (CONN_STATE_HTTP2)
- * - Header parsing timeout (CONN_STATE_READING_REQUEST with partial data)
- * - Global connection lifetime limit (defense-in-depth)
- *
- * Returns: 1 if timed out (connection closed), 0 otherwise
- */
-int
-server_check_connection_timeout (SocketHTTPServer_T server,
-                                 ServerConnection *conn,
-                                 int64_t now)
-{
-  if (check_global_lifetime_timeout (server, conn, now))
-    return 1;
-  if (check_tls_handshake_timeout (server, conn, now))
-    return 1;
-  if (check_request_timeout (server, conn, now))
-    return 1;
-  if (check_response_timeout (server, conn, now))
-    return 1;
-  if (check_http2_timeout (server, conn, now))
-    return 1;
-
-  return 0;
 }
 
 SocketHTTP_Method
@@ -1411,92 +756,6 @@ SocketHTTPServer_Request_upgrade_websocket (
   END_TRY;
 
   return NULL; /* Only reached on alloc failures before accept */
-}
-
-/**
- * @brief Validate RFC 8441 Extended CONNECT for WebSocket over HTTP/2.
- *
- * Checks that the request is a valid WebSocket upgrade per RFC 8441:
- * - Method must be CONNECT
- * - :protocol pseudo-header must be "websocket"
- * - Sec-WebSocket-Version must be "13" if present
- *
- * @param s HTTP/2 stream to validate
- * @return 1 if valid WebSocket upgrade, 0 otherwise
- */
-static int
-validate_rfc8441_websocket_upgrade (ServerHTTP2Stream *s)
-{
-  const char *version;
-
-  /* RFC 8441 Extended CONNECT: :method=CONNECT, :protocol=websocket */
-  if (s->request->method != HTTP_METHOD_CONNECT)
-    return 0;
-  if (s->h2_protocol == NULL || strcmp (s->h2_protocol, "websocket") != 0)
-    return 0;
-
-  /* Validate WebSocket version if specified */
-  version
-      = SocketHTTP_Headers_get (s->request->headers, "Sec-WebSocket-Version");
-  if (version != NULL && strcmp (version, "13") != 0)
-    return 0;
-
-  return 1;
-}
-
-/**
- * @brief Prepare HTTP/2 200 response for WebSocket upgrade.
- *
- * Allocates response headers if needed and populates the response structure.
- *
- * @param req Request context (provides arena for allocation)
- * @param s HTTP/2 stream
- * @param response Response structure to populate
- * @return 0 on success, -1 on failure (headers already sent or alloc failed)
- */
-static int
-prepare_h2_websocket_response (SocketHTTPServer_Request_T req,
-                               ServerHTTP2Stream *s,
-                               SocketHTTP_Response *response)
-{
-  if (s->response_headers_sent)
-    return -1;
-
-  if (s->response_headers == NULL)
-    s->response_headers = SocketHTTP_Headers_new (req->arena);
-  if (s->response_headers == NULL)
-    return -1;
-
-  s->response_status = 200;
-
-  memset (response, 0, sizeof (*response));
-  response->version = HTTP_VERSION_2;
-  response->status_code = 200;
-  response->headers = s->response_headers;
-
-  return 0;
-}
-
-/**
- * @brief Configure stream for WebSocket-over-HTTP/2 data delivery.
- *
- * Sets streaming flags and registers the callback for DATA frame delivery.
- *
- * @param s HTTP/2 stream to configure
- * @param callback Callback for incoming WebSocket frame data
- * @param userdata User context passed to callback
- */
-static void
-setup_ws_over_h2_streaming (ServerHTTP2Stream *s,
-                            SocketHTTPServer_BodyCallback callback,
-                            void *userdata)
-{
-  s->response_streaming = 1;
-  s->response_headers_sent = 1;
-  s->body_streaming = 1;
-  s->body_callback = callback;
-  s->body_callback_userdata = userdata;
-  s->ws_over_h2 = 1;
 }
 
 SocketHTTP2_Stream_T
